# src/lib/api-client/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Prophezeiung API",
    "version": "1.0.0",
    "description": "API for the Prophezeiung prediction platform"
  },
  "servers": [
    {
      "url": "http://localhost:3001",
      "description": "Development"
    }
  ],
  "security": [
    {
      "cookieAuth": []
    }
  ],
  "components": {
    "securitySchemes": {
      "cookieAuth": {
        "type": "apiKey",
        "in": "cookie",
        "name": "session",
        "description": "Session cookie authentication using JWT"
      }
    },
    "schemas": {
...
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "error": {
            "type": "string",
            "description": "Error message"
          }
        },
        "required": ["error"]
      },
...
      "PasswordLoginStatus": {
        "type": "object",
        "properties": {
          "passwordLoginEnabled": {
            "type": "boolean"
          },
          "forcePasswordChange": {
            "type": "boolean"
          },
          "hasPasskeys": {
            "type": "boolean"
          },
          "canDisablePasswordLogin": {
            "type": "boolean"
          }
        },
        "required": [
          "passwordLoginEnabled",
          "forcePasswordChange",
          "hasPasskeys",
          "canDisablePasswordLogin"
        ]
      },
      "TogglePasswordLoginResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean"
          },
          "passwordLoginEnabled": {
            "type": "boolean"
          },
          "message": {
            "type": "string"
          }
        },
        "required": ["success", "passwordLoginEnabled", "message"]
      },
      "TogglePasswordLoginRequest": {
        "type": "object",
        "properties": {
          "enabled": {
            "type": "boolean"
          }
        },
        "required": ["enabled"]
      },
...
    },
...
  },
  "paths": {
...
    "/api/users/me/password-login": {
      "get": {
        "tags": ["User"],
        "summary": "Get password login status",
        "responses": {
          "200": {
            "description": "Password login status",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PasswordLoginStatus"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": ["User"],
        "summary": "Toggle password login",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TogglePasswordLoginRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Password login toggled",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TogglePasswordLoginResponse"
                }
              }
            }
          },
          "400": {
            "description": "Cannot disable without passkey",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
  }
}